Move setRuntimeMetadata to new file
