+++ node-src/tasks/initialize/index.test.ts
@@ -1,94 +1,19 @@
-import { getCliCommand as getCliCommandDefault } from '@antfu/ni';
 import { validateStorybookReactNativeVersion as validateStorybookReactNativeVersionDefault } from '@cli/react-native/validateStorybookVersion';
 import TestLogger from '@cli/testLogger';
-import { execa as execaDefault } from 'execa';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { announceBuild, setRuntimeMetadata } from './index';
+import { announceBuild } from './index';
 
-vi.mock('@antfu/ni');
-vi.mock('execa', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('execa')>();
-  return {
-    ...actual,
-    execa: vi.fn(() => Promise.resolve()),
-  };
-});
 vi.mock('../../lib/react-native/validateStorybookVersion', () => ({
   validateStorybookReactNativeVersion: vi.fn().mockResolvedValue(undefined),
 }));
 
-const execa = vi.mocked(execaDefault);
-const getCliCommand = vi.mocked(getCliCommandDefault);
 const validateStorybookReactNativeVersion = vi.mocked(validateStorybookReactNativeVersionDefault);
 
 const log = new TestLogger();
 
 beforeEach(() => {
   vi.clearAllMocks();
-});
-
-describe('setRuntimeMetadata', () => {
-  beforeEach(() => {
-    execa.mockReturnValue(Promise.resolve({ stdout: '1.2.3' }) as any);
-  });
-
-  it('sets the build command on the context', async () => {
-    getCliCommand.mockReturnValue(Promise.resolve('npm'));
-
-    const ctx = {
-      sourceDir: './source-dir/',
-      options: { buildScriptName: 'build:storybook' },
-      storybook: { version: '6.2.0' },
-      git: { changedFiles: ['./index.js'] },
-    } as any;
-    await setRuntimeMetadata(ctx);
-
-    expect(ctx.runtimeMetadata).toEqual({
-      nodePlatform: expect.stringMatching(/darwin|linux|win32/),
-      nodeVersion: process.versions.node,
-      packageManager: 'npm',
-      packageManagerVersion: '1.2.3',
-    });
-  });
-
-  it('supports yarn', async () => {
-    getCliCommand.mockReturnValue(Promise.resolve('yarn'));
-
-    const ctx = {
-      sourceDir: './source-dir/',
-      options: { buildScriptName: 'build:storybook' },
-      storybook: { version: '6.1.0' },
-      git: {},
-    } as any;
-    await setRuntimeMetadata(ctx);
-
-    expect(ctx.runtimeMetadata).toEqual({
-      nodePlatform: expect.stringMatching(/darwin|linux|win32/),
-      nodeVersion: process.versions.node,
-      packageManager: 'yarn',
-      packageManagerVersion: '1.2.3',
-    });
-  });
-
-  it('supports pnpm', async () => {
-    getCliCommand.mockReturnValue(Promise.resolve('pnpm'));
-
-    const ctx = {
-      sourceDir: './source-dir/',
-      options: { buildScriptName: 'build:storybook' },
-      storybook: { version: '6.1.0' },
-      git: {},
-    } as any;
-    await setRuntimeMetadata(ctx);
-
-    expect(ctx.runtimeMetadata).toEqual({
-      nodePlatform: expect.stringMatching(/darwin|linux|win32/),
-      nodeVersion: process.versions.node,
-      packageManager: 'pnpm',
-      packageManagerVersion: '1.2.3',
-    });
-  });
 });
 
 describe('announceBuild', () => {
